test: simplify unit testing design
Previoulsy I tried to gather all unit tests of a submodule into a single
executable using 'create_test_sourcelist' function, but because of its
limitations it resulted in an over-engineered design.

We now have one executable per unit test, and we will deal with the
issue this can bring if we ever find any.